requirements.txt
google-cloud-storage>=2.7.0
google-cloud-bigquery>=3.4.0
pandas>=1.5.0
pytest
fsspec
gcsfs